test: move all tests in tests module
