+++ bdk-ffi/src/tests/mod.rs
@@ -1,0 +1,5 @@
+mod bitcoin;
+mod descriptor;
+mod error;
+mod keys;
+mod tx_builder;
